+++ math/m_sparse.h
@@ -745,7 +745,7 @@ if (NOT(condition))			\
         if (( (op1) > LARGEST_SHORT_INTEGER AND (op2) != 0) OR \
             ( (op2) > LARGEST_SHORT_INTEGER AND (op1) != 0)) \
         {   double fProduct = (double)(op1) * (double)(op2); \
-            if (fProduct >= LARGEST_LONG_INTEGER) \
+            if (fProduct >= (double)LARGEST_LONG_INTEGER) \
                 (product) = LARGEST_LONG_INTEGER; \
             else \
                 (product) = (long)fProduct; \
